Add unified R_ShipController
